fix: integration security and quality issues, add Discord/Slack docs
- Enforce SSRF check at config validation time for generic webhooks
- Consolidate SSRF logic into config.ts, remove duplication from client.ts
- Mask webhook URL in generic webhook API response (consistent with Discord/Slack)
- Fix double body read in github import-issues middleware
- Fix Slack frontend useEffect deps to match Discord's stable reset pattern
- Add Slack-specific URL pattern validation on frontend
- Add $onUpdate to githubIntegrationTable.updatedAt
- Add dedicated Discord and Slack integration doc pages
