+++ apps/api/src/database/schema.ts
@@ -445,7 +445,10 @@ export const githubIntegrationTable = pgTable("github_integration", {
   installationId: integer("installation_id"),
   isActive: boolean("is_active").default(true),
   createdAt: timestamp("created_at", { mode: "date" }).defaultNow().notNull(),
-  updatedAt: timestamp("updated_at", { mode: "date" }).defaultNow().notNull(),
+  updatedAt: timestamp("updated_at", { mode: "date" })
+    .defaultNow()
+    .$onUpdate(() => new Date())
+    .notNull(),
 });
 
 export const integrationTable = pgTable(
