+++ apps/docs/core/integrations/slack.mdx
@@ -1,0 +1,68 @@
+---
+title: Slack
+description: Send task notifications from Kaneo to a Slack channel using incoming webhooks.
+---
+
+Kaneo can post task activity to a Slack channel whenever events occur in a project. This uses Slack's incoming webhook feature — no bot token or OAuth flow required.
+
+## Create a Slack webhook
+
+1. Go to [api.slack.com/apps](https://api.slack.com/apps) and click **Create New App** then **From scratch**.
+2. Name your app (e.g. "Kaneo") and select your workspace.
+3. In the app settings, go to **Incoming Webhooks** and toggle it **On**.
+4. Click **Add New Webhook to Workspace**.
+5. Select the channel where you want notifications (or choose your own DM for personal notifications).
+6. Click **Allow**, then copy the webhook URL. It will look like:
+
+```
+https://hooks.slack.com/services/T00000000/B00000000/XXXXXXXXXXXXXXXXXXXXXXXX
+```
+
+<Warning>
+If your Slack workspace restricts app installations, you may need a workspace admin to approve the app before you can use the webhook.
+</Warning>
+
+## Connect Slack in Kaneo
+
+1. Open **Settings** in the sidebar.
+2. Under **Projects**, select the project you want to connect.
+3. Open the **Integrations** tab.
+4. Expand the **Slack** section.
+5. Paste your webhook URL.
+6. Optionally add a channel name label for your own reference.
+7. Choose which events to subscribe to (see below).
+8. Click **Connect**.
+
+Once connected you can enable or disable the integration with the toggle, update event subscriptions, or disconnect entirely.
+
+## Supported events
+
+| Event | Default | Description |
+| --- | --- | --- |
+| Task created | On | A new task is created in the project |
+| Task status changed | On | A task moves to a different status column |
+| Task priority changed | Off | A task's priority is changed |
+| Task title changed | Off | A task's title is edited |
+| Task description changed | Off | A task's description is edited |
+| Task comment created | On | A new comment is added to a task |
+
+You can toggle any combination of these events in the integration settings at any time.
+
+## Message format
+
+Kaneo sends Slack Block Kit messages with:
+
+- A header describing the event (e.g. "Task Status Changed")
+- Task name with a clickable link to the task in Kaneo
+- Fields for project, status, and priority
+- The name of the user who triggered the event
+
+Make sure [`KANEO_CLIENT_URL`](/core/installation/environment-variables) is set to your public Kaneo URL so task links work correctly.
+
+## Updating the webhook
+
+To change the webhook URL after connecting, paste the new URL in the webhook field and click **Save Changes**. If you leave the field empty, the existing URL is kept.
+
+## Disconnecting
+
+Click **Disconnect** to remove the integration. This deletes the stored webhook URL and all event settings for that project. You can reconnect at any time by pasting a new webhook URL.
