+++ apps/docs/docs.json
@@ -80,8 +80,12 @@
 						]
 					},
 					{
-						"group": "Outgoing webhooks",
-						"pages": ["core/integrations/outgoing-webhooks"]
+						"group": "Integrations",
+						"pages": [
+							"core/integrations/discord",
+							"core/integrations/slack",
+							"core/integrations/outgoing-webhooks"
+						]
 					},
 					{
 						"group": "Authentication Providers",
